control exactly where load instructions are emitted.
For memories we need to load the address and data element at the address.
Before loads were pushed onto the front of the 1st basic block which
reversed the address/data load and caused an assert.  We now split
each funtion into 2 basic blocks (which should be collapsed) - one
for the loads, one for the operations.  As loads are generated they
are pushed onto the end on the loading basic block (just before the
jump) so everything is in order.

/rant on/ the irbuilder interface is *seriously* horrible.